# perf: avoid closures that cause allocations