--- a/.gitignore
+++ b/.gitignore
@@ -12,3 +12,5 @@
 *.out
 
 /go.work*
+
+*.bench